Add feature to include current route slug in body
This pull requests extends Upstament's Router by allowing users whether they'd like to display the sanitized slug of the current route in the <body> tag or not. This may play a big role in CSS and other areas.

Addionally, one parameter has been added to the `map()` funciton, after `args`, called `add_to_body` which is boolean and set to `true` on default. However, this can be changed invidually for any loaded route fike.

Signed-off-by: Kolja Nolte <kolja.nolte@gmail.com>